+++ sakkada/admin/tree_list/__init__.py
@@ -5,7 +5,7 @@ from django.core import checks
 from django.forms.widgets import Media
 from django.http import JsonResponse
 from django.utils.safestring import mark_safe
-from django.utils.translation import ugettext_lazy as _
+from django.utils.translation import gettext_lazy as _
 
 
 class BaseTreeAdmin(admin.ModelAdmin):