+++ sakkada/admin/views/cache_clear/__init__.py
@@ -1,5 +1,5 @@
 from django.contrib import messages
-from django.utils.translation import ugettext_lazy as _
+from django.utils.translation import gettext_lazy as _
 from django.shortcuts import render, redirect
 from django.core.cache import caches