+++ sakkada/system/files/storage/unique.py
@@ -9,7 +9,7 @@ Changes:
 
 from django.core.files import File
 from django.core.files.storage import FileSystemStorage
-from django.utils.encoding import force_text
+from django.utils.encoding import force_str
 
 
 class NoAvailableName(Exception):
@@ -71,7 +71,7 @@ class UniqueNameFileSystemStorage(FileSystemStorage):
         name = self._save(name, content)
 
         # Store filenames with forward slashes, even on Windows
-        return force_text(name.replace('\\', '/'))
+        return force_str(name.replace('\\', '/'))
 
     def _save(self, name, content):
         # Save original file without changes if required