+++ sakkada/system/validators.py
@@ -1,5 +1,5 @@
 import os
-from django.utils.translation import ugettext_lazy as _
+from django.utils.translation import gettext_lazy as _
 from django.utils.deconstruct import deconstructible
 from django.core.exceptions import ValidationError
 from django.db.models.fields.files import FieldFile