+++ CHANGES.md
@@ -1,6 +1,12 @@
 Changes history
 ===============
 
+3.0.0   (2020-01-09)
+--------------------
+-   Changed python and django supported versions to 3.6.x and 3.0.x respectively.
+-   Added tests for multivalue fields and prev_next models.
+-   Removed simplefb and tinymce editors.
+
 2.0.0   (2018-07-10)
 --------------------
 -   Changed python and django supported versions to 3.6.x and 2.0.x respectively.